Refactors limb script
