+++ UnityProject/Assets/Prefabs/Items/Implants/Organs/HumanRightLeg.prefab
@@ -349,7 +349,8 @@ PrefabInstance:
       value: 
       objectReference: {fileID: 11400000, guid: f0594d5820b2ebd4d90a67a4b81c7ba0,
         type: 2}
-    m_RemovedComponents: []
+    m_RemovedComponents:
+    - {fileID: 3959904421835429671, guid: 6a1a4888888e8994187da2ccb8d6f1c2, type: 3}
   m_SourcePrefab: {fileID: 100100000, guid: 6a1a4888888e8994187da2ccb8d6f1c2, type: 3}
 --- !u!1 &6498312071342856297 stripped
 GameObject:
@@ -357,6 +358,23 @@ GameObject:
     type: 3}
   m_PrefabInstance: {fileID: 1050759302020462618}
   m_PrefabAsset: {fileID: 0}
+--- !u!114 &-3757956373148372916
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6498312071342856297}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 73c8a0e70c3343fd9c7d764c28769377, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  isEMPVunerable: 0
+  EMPResistance: 2
+  walkingSpeed: 3
+  runningSpeed: 4.5
+  legEfficiency: 1
 --- !u!114 &5655852655570756341
 MonoBehaviour:
   m_ObjectHideFlags: 0
