+++ UnityProject/Assets/Scripts/Items/Implants/Limbs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 17f4cf4ba47798c40b075c8277d32240
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
